requirements.txt
torch~=1.8.1
numpy~=1.18.1
tqdm~=4.32.1
setuptools~=46.0.0
yfinance~=0.1.63
gin-config~=0.4.0
pandas~=1.2.2
torchviz~=0.0.2